Rename Asset.js to InternalAsset.js
