+++ packages/core/core/src/InternalAsset.js
@@ -33,14 +33,9 @@ type AssetOptions = {|
   id?: string,
   hash?: ?string,
   idBase?: ?string,
-  // cache: Cache,
-  // fs: FileSystem,
   filePath: FilePath,
   type: string,
-  // content?: Blob,
   contentKey?: ?string,
-  // ast?: ?AST,
-  // map?: ?SourceMap,
   mapKey?: ?string,
   dependencies?: Map<string, Dependency>,
   connectedFiles?: Map<FilePath, File>,
